refactor: move stories to categorized folders
