+++ stories/molecules/Header.stories.js
@@ -3,7 +3,6 @@ import React from 'react';
 import { Header } from './Header';
 
 export default {
-  title: 'Example/Header',
   component: Header,
   play: async () => {},
 };
