+++ stories/pages/Page.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import { Header } from './Header';
+import { Header } from '../molecules/Header';
 import './page.css';
 
 export const Page = ({ user, onLogin, onLogout, onCreateAccount }) => (
